+++ CHANGELOG.md
@@ -67,6 +67,7 @@
 - Fix incorrect session refreshing from the database !4109 @DL6ER
 - Fix misleading link in the bell for suggested food-share points #2370 !4112 #irgendwer
 - Fix length restriction when creating a store #2338 !4050 @skurka
+- Fix incorrect confirmation dialogue #2380 !4165 @DL6ER
 
 ## Refactoring
 - Get rid of xhr completely !3826 !3933 @chriswalg